Create list of characters home screen

+++ app/src/main/java/com/marvel/characters/model/CharactersDataModel.kt
@@ -6,7 +6,7 @@ class CharactersDataModel : Serializable {
     lateinit var results: ArrayList<CharacterModel>
     var offset: Int = 0
     var limit: Int = 0
-    var total: Long = 0
+    var total: Int = 0
     var count: Int = 0
 
 }
